app/requirements.txt
###### Requirements without Version Specifiers ######

###### Requirements with Version Specifiers ######
#   See https://www.python.org/dev/peps/pep-0440/#version-specifiers

certifi==2019.11.28
chardet==3.0.4
gspread==3.2.0
httplib2==0.18.0
idna==2.9
oauth2client==4.1.3
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.11.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
urllib3==1.25.8

###### Refer to other requirements files ######
# -r other-requirements.txt
#
#

###### A particular file ######
# ./downloads/numpy-1.9.2-cp34-none-win32.whl
# http://wxpython.org/Phoenix/snapshot-builds/wxPython_Phoenix-3.0.3.dev1820+49a8884-cp34-none-win_amd64.whl
